+++ pkg/scalers/aws_sqs_queue_scaler_test.go
@@ -207,6 +207,22 @@ var testAWSSQSMetadata = []parseAWSSQSMetadataTestData{
 		testAWSSQSAuthentication,
 		false,
 		"properly formed queue and region"},
+	{map[string]string{
+		"queueURL":    testAWSSimpleQueueURL,
+		"queueLength": "1",
+		"awsRegion":   "eu-west-1",
+		"scaleOnInFlight": false},
+		testAWSSQSAuthentication,
+		false,
+		"properly formed queue and region"},
+		{map[string]string{
+			"queueURL":    testAWSSimpleQueueURL,
+			"queueLength": "1",
+			"awsRegion":   "eu-west-1",
+			"scaleOnInFlight": true},
+			testAWSSQSAuthentication,
+			false,
+			"properly formed queue and region"},
 }
 
 var awsSQSMetricIdentifiers = []awsSQSMetricIdentifier{